Add kernel rootkit KPM native interface implementation
This commit introduces a new source file `rootkit_kpm_native.cpp` that implements the kernel rootkit KPM (Kernel Module Payload) interface. The file includes functions for reading KPM files, executing KPM commands, loading and unloading KPM modules, and listing loaded modules. It utilizes system calls and dmesg output to interact with the kernel, ensuring proper error handling and validation throughout the process.

+++ Lite_version/src/patch_kernel_root/patch_audit_log_start.cpp
@@ -5,7 +5,7 @@ using namespace asmjit;
 using namespace asmjit::a64;
 using namespace asmjit::a64::Predicate;
 
-PatchAuditLogStart::PatchAuditLogStart(const PatchBase& patch_base, uint64_t audit_log_start) : PatchBase(patch_base) {
+PatchAuditLogStart::PatchAuditLogStart(const PatchBase& patch_base, size_t audit_log_start) : PatchBase(patch_base) {
 	m_audit_log_start_orig_entry_insn = *(uint32_t*)&m_file_buf[audit_log_start];
 	m_audit_log_start = skip_pac_bti_at_func_start(audit_log_start);
 }
